Restructure KeyBit file tree
