+++ src/components/MapCanvas/WorldManager/FrameHandler/ImageLoader/index.ts
